+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - [bb tasks] Add support for docker `:network` option
 - [docs] Add FAQ, CI support documents
 - [bb tasks] Fix using docker in CI environment
+- [bb tasks] Remove broken stack:lint command. Use plain clj-kondo instead
 
 ## 0.2.1 (10-06-2021)